+++ backend/src/identity/provider-type.enum.ts
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
@@ -11,8 +11,5 @@ export enum ProviderType {
   OAUTH2 = 'oauth2',
   GITLAB = 'gitlab',
   GITHUB = 'github',
-  FACEBOOK = 'facebook',
-  TWITTER = 'twitter',
-  DROPBOX = 'dropbox',
   GOOGLE = 'google',
 }