+++ frontend/src/api/config/types.ts
@@ -24,11 +24,8 @@ export enum GuestAccessLevel {
 }
 
 export enum AuthProviderType {
-  DROPBOX = 'dropbox',
-  FACEBOOK = 'facebook',
   GITHUB = 'github',
   GOOGLE = 'google',
-  TWITTER = 'twitter',
   GITLAB = 'gitlab',
   OAUTH2 = 'oauth2',
   LDAP = 'ldap',
@@ -43,22 +40,16 @@ export type AuthProviderTypeWithCustomName =
   | AuthProviderType.SAML
 
 export type AuthProviderTypeWithoutCustomName =
-  | AuthProviderType.DROPBOX
-  | AuthProviderType.FACEBOOK
   | AuthProviderType.GITHUB
   | AuthProviderType.GOOGLE
-  | AuthProviderType.TWITTER
   | AuthProviderType.LOCAL
 
 export const authProviderTypeOneClick = [
-  AuthProviderType.DROPBOX,
-  AuthProviderType.FACEBOOK,
   AuthProviderType.GITHUB,
   AuthProviderType.GITLAB,
   AuthProviderType.GOOGLE,
   AuthProviderType.OAUTH2,
-  AuthProviderType.SAML,
-  AuthProviderType.TWITTER
+  AuthProviderType.SAML
 ]
 
 export interface AuthProviderWithCustomName {